refactor(web): dissolve domains/avatar into shared top-level directories (LUM-1917)
The avatar domain has no routes or pages — it is shared infrastructure
consumed by 5+ domains (chat, home, onboarding, intelligence) and
top-level components. Per CONVENTIONS.md, code used by two or more
domains belongs in top-level shared directories.

Redistribution:
- types.ts → types/avatar.ts
- api.ts → utils/avatar-api.ts
- svg-compositor.ts → utils/avatar-svg-compositor.ts
- bundled-components.ts → utils/avatar-bundled-components.ts
- random.ts → utils/avatar-random.ts
- subagent-avatar.ts → utils/avatar-subagent.ts
- use-assistant-avatar.ts (+test) → hooks/use-assistant-avatar.ts
- use-dynamic-favicon.ts → hooks/use-dynamic-favicon.ts
- subagent-avatar-chip.tsx (+test) → components/avatar/

Removes 11 entries from .cross-domain-allowlist.json (all avatar
cross-domain violations eliminated). The domains/avatar/ directory
is deleted entirely.

Closes LUM-1917
