+++ apps/web/src/components/avatar/animated-avatar.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useMemo, useRef, useState } from "react";
 import { useReducedMotion } from "motion/react";
 
-import { computeTransforms, resolveDefinitions } from "@/domains/avatar/svg-compositor.js";
-import type { CharacterComponents, CharacterTraits, EyePathDefinition } from "@/domains/avatar/types.js";
+import { computeTransforms, resolveDefinitions } from "@/utils/avatar-svg-compositor.js";
+import type { CharacterComponents, CharacterTraits, EyePathDefinition } from "@/types/avatar.js";
 
 interface AnimatedAvatarProps {
   components: CharacterComponents;
