+++ apps/web/src/components/avatar/avatar-customization-panel.tsx
@@ -1,8 +1,8 @@
 import { ChevronLeft, ChevronRight, Dices, Save, X } from "lucide-react";
 import { useCallback, useEffect, useRef, useState } from "react";
 
-import { fetchCharacterComponents, saveCharacterTraits } from "@/domains/avatar/api.js";
-import type { CharacterComponents, CharacterTraits } from "@/domains/avatar/types.js";
+import { fetchCharacterComponents, saveCharacterTraits } from "@/utils/avatar-api.js";
+import type { CharacterComponents, CharacterTraits } from "@/types/avatar.js";
 import { AvatarRenderer } from "@/components/avatar-renderer.js";
 
 export interface AvatarCustomizationPanelProps {
