+++ apps/web/src/components/avatar/subagent-avatar-chip.test.tsx
@@ -1,8 +1,8 @@
 import { describe, expect, test } from "bun:test";
 import { renderToStaticMarkup } from "react-dom/server";
 
-import { SubagentAvatarChip } from "@/domains/avatar/subagent-avatar-chip.js";
-import { subagentTraits } from "@/domains/avatar/subagent-avatar.js";
+import { SubagentAvatarChip } from "@/components/avatar/subagent-avatar-chip.js";
+import { subagentTraits } from "@/utils/avatar-subagent.js";
 
 describe("SubagentAvatarChip", () => {
   test("is deterministic: same subagentId renders identical DOM", () => {
