+++ apps/web/src/domains/home/home-greeting-header.tsx
@@ -2,7 +2,7 @@ import { SquarePen } from "lucide-react";
 
 import { Button, Typography } from "@vellum/design-library";
 import { ChatAvatar } from "@/components/avatar/chat-avatar.js";
-import type { CharacterComponents, CharacterTraits } from "@/domains/avatar/types.js";
+import type { CharacterComponents, CharacterTraits } from "@/types/avatar.js";
 
 interface HomeGreetingHeaderProps {
   avatarComponents: CharacterComponents | null;
