+++ apps/web/src/domains/intelligence/components/constellation-view/constellation-view.tsx
@@ -20,7 +20,7 @@ import {
 
 import { Button } from "@vellum/design-library";
 import { inferCategory } from "@/domains/intelligence/skills/category.js";
-import type { CharacterComponents, CharacterTraits } from "@/domains/avatar/types.js";
+import type { CharacterComponents, CharacterTraits } from "@/types/avatar.js";
 import type { SkillInfo } from "@/domains/intelligence/skills/types.js";
 
 import {
