+++ apps/web/src/hooks/use-assistant-avatar.ts
@@ -5,8 +5,8 @@ import {
   fetchCharacterComponents,
   fetchCharacterTraits,
   fetchAvatarImageUrl,
-} from "./api.js";
-import type { CharacterComponents, CharacterTraits } from "./types.js";
+} from "@/utils/avatar-api.js";
+import type { CharacterComponents, CharacterTraits } from "@/types/avatar.js";
 import { avatarQueryKey } from "@/lib/sync/query-tags.js";
 
 interface AvatarData {
