+++ apps/web/src/types/avatar.ts
