+++ apps/web/src/utils/avatar-bundled-components.ts
@@ -6,7 +6,7 @@
  * Keep in sync when body shapes, eye styles, or colors change.
  */
 
-import type { CharacterComponents } from "@/domains/avatar/types.js";
+import type { CharacterComponents } from "@/types/avatar.js";
 
 const SCLERA = "#F2F2F2";
 const PUPIL = "#1A1A1A";
